refactor: Consolidate Ollama functionality into dedicated modules and remove legacy service layer
