+++ src/scanning_tool/ollama/installer.py
@@ -1,0 +1,156 @@
+import logging
+import os
+import shutil
+import subprocess
+import sys
+import webbrowser
+from typing import Optional
+
+from .host import get_ollama_host, is_local_ollama_host
+
+logger = logging.getLogger(__name__)
+
+
+def show_installation_message(system_name: str) -> None:
+    """Present a final installation message, using a GUI prompt on Windows."""
+    import tkinter as tk
+    import tkinter.messagebox as messagebox
+
+    message = (
+        f"Ollama installation initiated for {system_name.title()}.\n\n"
+        "After installation completes:\n"
+        "1. Restart this program\n"
+        "2. The first scan will download the AI model automatically\n\n"
+        "Visit https://ollama.com/ for troubleshooting."
+    )
+
+    if system_name == "windows":
+        temp_root = None
+        try:
+            temp_root = tk.Tk()
+            temp_root.withdraw()
+            messagebox.showinfo("Ollama Installation", message, parent=temp_root)
+        except Exception as exc:
+            logger.debug(f"Unable to show Windows message box: {exc}")
+            logger.info(message)
+        else:
+            logger.info(message)
+        finally:
+            if temp_root is not None:
+                temp_root.destroy()
+    else:
+        logger.info(message)
+
+
+def ensure_ollama_installed() -> None:
+    """Check whether Ollama is installed locally when required."""
+    host = get_ollama_host()
+    if not is_local_ollama_host(host):
+        logger.info(f"Using remote Ollama host at {host}; skipping local installation check.")
+        return
+
+    if not shutil.which("ollama"):
+        import platform
+
+        system = platform.system().lower()
+
+        logger.info("Ollama not found on your system.")
+        logger.info("Ollama is required for AI-powered code recognition.")
+        logger.info("")
+
+        if system == "windows":
+            logger.info("=== Windows Installation Options ===")
+            logger.info("1. Automatic download and install (Recommended)")
+            logger.info("2. Manual download from website")
+            logger.info("")
+
+            download_url = "https://ollama.com/download/OllamaSetup.exe"
+            logger.info("Opening the Ollama download link in your default browser...")
+            logger.info(f"Download URL: {download_url}")
+            try:
+                opened = webbrowser.open(download_url)
+                if opened:
+                    logger.info("Browser opened successfully. Follow the prompts to install Ollama.")
+                else:
+                    logger.warning("The browser did not report success. Please open the link manually if nothing happens.")
+            except Exception as e:
+                logger.error(f"Unable to open browser automatically: {e}")
+                logger.info("Please open the link manually to download Ollama.")
+
+        elif system == "linux":
+            logger.info("=== Linux Installation Options ===")
+
+            distro_info = ""
+            package_cmd = ""
+
+            try:
+                with open("/etc/os-release", "r") as f:
+                    os_release = f.read().lower()
+
+                if "debian" in os_release or "ubuntu" in os_release or "mint" in os_release:
+                    distro_info = "Debian/Ubuntu/Mint"
+                    package_cmd = "curl -fsSL https://ollama.com/install.sh | sh"
+                elif "arch" in os_release or "manjaro" in os_release:
+                    distro_info = "Arch/Manjaro"
+                    package_cmd = "sudo pacman -S ollama"
+                elif "fedora" in os_release or "rhel" in os_release or "centos" in os_release:
+                    distro_info = "RedHat/Fedora/CentOS"
+                    package_cmd = "curl -fsSL https://ollama.com/install.sh | sh"
+                elif "gentoo" in os_release or "funtoo" in os_release:
+                    distro_info = "Gentoo/Funtoo"
+                    package_cmd = "sudo emerge --ask ollama"
+                elif "suse" in os_release or "opensuse" in os_release:
+                    distro_info = "SUSE/openSUSE"
+                    package_cmd = "sudo zypper install ollama"
+                else:
+                    distro_info = "Unknown Linux"
+                    package_cmd = "curl -fsSL https://ollama.com/install.sh | sh"
+            except Exception:
+                distro_info = "Unknown Linux"
+                package_cmd = "curl -fsSL https://ollama.com/install.sh | sh"
+
+            logger.info(f"Detected: {distro_info}")
+            logger.info(f"Recommended command: {package_cmd}")
+            logger.info("")
+            logger.info("1. Run the recommended installation command")
+            logger.info("2. Manual installation from website")
+            logger.info("")
+
+            choice = input("Would you like to run the installation command? (y/n): ").lower().strip()
+
+            if choice in ["y", "yes", "1", ""]:
+                logger.info(f"Running: {package_cmd}")
+                logger.info("Please enter your password if prompted...")
+                try:
+                    result = subprocess.run(package_cmd, shell=True, check=False)
+                    if result.returncode == 0:
+                        logger.info("Ollama installation completed!")
+                        logger.info("Please restart this program to continue.")
+                    else:
+                        logger.warning("Installation failed or was cancelled.")
+                        logger.info("You can try installing manually from https://ollama.com/")
+                except Exception as e:
+                    logger.error(f"Error running installation command: {e}")
+                    logger.info("Please visit https://ollama.com/ for manual installation.")
+            else:
+                logger.info("Opening Ollama website for manual installation...")
+                webbrowser.open("https://ollama.com/")
+
+        else:
+            logger.info("=== Unsupported Operating System ===")
+            logger.info("This tool currently supports Windows and Linux only.")
+            logger.info("Please install Ollama manually from: https://ollama.com/")
+            webbrowser.open("https://ollama.com/")
+
+        show_installation_message(system)
+
+        input("\nPress ENTER after installing Ollama to close this program...")
+        sys.exit(0)
+
+    else:
+        try:
+            version = subprocess.check_output(["ollama", "--version"], text=True).strip()
+            logger.info(f"Ollama found: {version}")
+        except Exception as e:
+            logger.error(f"Error checking Ollama: {e}")
+            sys.exit("Please install Ollama and rerun this program.")
